+++ test/data.jl
@@ -3,6 +3,7 @@
     Y = [1:5;]
 
     d = DataLoader(X, batchsize=2)
+    @inferred first(d)
     batches = collect(d)
     @test eltype(batches) == eltype(d) == typeof(X)
     @test length(batches) == 3
@@ -11,6 +12,7 @@
     @test batches[3] == X[:,5:5]
 
     d = DataLoader(X, batchsize=2, partial=false)
+    @inferred first(d)
     batches = collect(d)
     @test eltype(batches) == eltype(d) == typeof(X)
     @test length(batches) == 2
@@ -18,6 +20,7 @@
     @test batches[2] == X[:,3:4]
 
     d = DataLoader((X,), batchsize=2, partial=false)
+    @inferred first(d)
     batches = collect(d)
     @test eltype(batches) == eltype(d) == Tuple{typeof(X)}
     @test length(batches) == 2
@@ -25,6 +28,7 @@
     @test batches[2] == (X[:,3:4],)
 
     d = DataLoader((X, Y), batchsize=2)
+    @inferred first(d)
     batches = collect(d)
     @test eltype(batches) == eltype(d) == Tuple{typeof(X), typeof(Y)}
     @test length(batches) == 3
@@ -37,6 +41,22 @@
     @test batches[2][2] == Y[3:4]
     @test batches[3][1] == X[:,5:5]
     @test batches[3][2] == Y[5:5]
+
+    # test with NamedTuple
+    d = DataLoader((x=X, y=Y), batchsize=2)
+    @inferred first(d)
+    batches = collect(d)
+    @test eltype(batches) == eltype(d) == NamedTuple{(:x, :y), Tuple{typeof(X), typeof(Y)}}
+    @test length(batches) == 3
+    @test length(batches[1]) == 2
+    @test length(batches[2]) == 2
+    @test length(batches[3]) == 2
+    @test batches[1][1] == batches[1].x == X[:,1:2]
+    @test batches[1][2] == batches[1].y == Y[1:2]
+    @test batches[2][1] == batches[2].x == X[:,3:4]
+    @test batches[2][2] == batches[2].y == Y[3:4]
+    @test batches[3][1] == batches[3].x == X[:,5:5]
+    @test batches[3][2] == batches[3].y == Y[5:5]
 
     # test interaction with `train!`
     θ = ones(2)